Created Processing context
close #113
