+++ pypro/modulos/tests/test_facade.py
@@ -1,13 +1,4 @@
-import pytest
-from model_bakery import baker
-
 from pypro.modulos import facade
-from pypro.modulos.models import Modulo
-
-
-@pytest.fixture
-def videos(db):
-    return [baker.make(Modulo, titulo=s) for s in 'Antes Depois'.split()]
 
 
 def test_listar_modulos_ordenados(modulos):
